==> Code/retrosynthesis/06-1.get_YeastMetaTwin.ipynb <==
import pandas as pd
import numpy as np
import json
import pandas as pd
import re
from rdkit import Chem
from rdkit.Chem import AllChem
from collections import Counter,defaultdict
from ast import literal_eval
import cobra
from cobra import Metabolite,Reaction,Gene
from tqdm import tqdm
import random
import pickle
from multiprocessing import Pool

import sys
# sys.path.append('../Script/')
from common import *
# from check_newinfo_getyeast8U import *

### input
yeast870_path =  '../../Data/model/yeast-GEM902.yml'
cut_off = 0.3
num = 50
num_process = 50
cut_off_path = f'../../Results/not_lipid/top{num}_{cut_off}_re/'
metabolites_info_to_GEM_path = cut_off_path + f'metabolites_info_to_GEM_top{num}_{cut_off}.csv'
yeast8_reaction_in_rxndb_json = cut_off_path + f'yeast8_reaction_in_rxndb_top{num}_{cut_off}.json'

filter_gene = 1  # default=1  #compound-protein score to filter gene,
#######################################################################################################################
filter_metabolite = 0  # 0.2/0.4   # default=0  #Randomly remove metabolites for revision
# 跑一遍0拿主要结果
# 最后跑一遍0.2 0.4看鲁棒性
#######################################################################################################################
if not filter_gene and not filter_metabolite:
    rxndb_total_info_to_model_path = cut_off_path + f'rxndb_total_info_to_model_top{num}_{cut_off}.csv'
elif filter_gene and not filter_metabolite:
    rxndb_total_info_to_model_path = cut_off_path + f'rxndb_total_info_to_model_top{num}_{cut_off}_espfiltergene.csv'
elif filter_gene and filter_metabolite:
    proportion = filter_metabolite
    rxndb_total_info_to_model_path = cut_off_path + f'rxndb_total_info_to_model_top{num}_{cut_off}_espfiltergene_randomfiltermet_{proportion}.csv' ####### random filter met
else:
    print('error')

### output
rxndb_total_info_to_model_del_reaction_path = '../../Results/not_lipid/top50_0.3_re/rxndb_total_info_to_model_del_reaction_top50_0.3.csv'
error_reaction_path = '../../Results/not_lipid/top50_0.3_re/error_reaction_random/'
error_reaction_all_path = '../../Results/not_lipid/top50_0.3_re/error_reaction_all.pickle'
yeast8U_del_path = '../../Data/model/yeast8U_del_re.yml'
yeast8U_mod_path = '../../Data/model/yeast8U_mod_re.yml'
biomass_up_reaction_path = '../../Results/not_lipid/top50_0.3_re/biomass_up_reaction_top50_0.3.pickle'

rxndb_total_info_to_model = pd.read_csv(rxndb_total_info_to_model_path)
rxndb_total_info_to_model.shape

get_yeast8U(metabolites_info_to_GEM_path,rxndb_total_info_to_model_path,yeast8_reaction_in_rxndb_json,yeast870_path,yeast8U_mod_path,metinfoID_2_YMDBID=None)

yeast8U_mod = cobra.io.load_yaml_model(yeast8U_mod_path)
yeast8U_mod.summary()

yeast8 = cobra.io.load_yaml_model(yeast870_path)


yeast8U_mod_path

yeast8U_mod

# Number of reactions 10618

# find reactions that affect biomass

reaction_id = []
for i in yeast8U_mod.reactions:
    if 'rxn' in i.id:
        i.bounds = (0,0)
        reaction_id.append(i.id)

biomass_up_reaction = []
for i in tqdm(reaction_id,total=len(reaction_id)):
    with yeast8U_mod:
        yeast8U_mod.reactions.get_by_id(i).bounds = (0,1000)
        yeast8U_mod.solver = 'gurobi'
        yeast8U_mod.optimize()
        if yeast8U_mod.reactions.get_by_id('r_2111').flux > 0.09:
            biomass_up_reaction.append(i)
            print(i,'error_reaction')
        else: 
            pass
        
biomass_up_reaction

# Save the list as a pickle file
with open(biomass_up_reaction_path, 'wb') as f:
    pickle.dump(biomass_up_reaction, f)


# # Load the pickle file
with open(biomass_up_reaction_path, 'rb') as f:
    biomass_up_reaction = pickle.load(f)

# View the loaded data
print(biomass_up_reaction)

atp_reaction = []

for reaction in yeast8U_mod.reactions:
    if 'rxn' in reaction.id:
        tmp = {str(k.id):v for k,v in reaction.metabolites.items()}
        if 's_0434' in tmp and tmp['s_0434']>0:
            atp_reaction.append(reaction.id)
            
print(len(atp_reaction))
print(atp_reaction)

atp_sim_name = ['UDP','UMP','UTP',
                'dUDP','dUMP','dUTP',
                'CDP','CMP','CTP',
                'dCDP','dCMP','dCTP',
                'TDP','TMP','TTP',
                'dTDP','dTMP','dTTP',
                'GDP','GMP','GTP',
                'dGDP','dGMP','dGTP',
                'IDP','IMP','ITP',
                'dIDP','dIMP','dITP',
                ]
atp_sim_list = []
for met in atp_sim_name:
    for i in yeast8.metabolites:
        if i.compartment == 'c':
            if i.name == met:
                print(i.id,i.name,met)
                atp_sim_list.append(i.id)

atp_sim_reaction = []

for reaction in yeast8U_mod.reactions:
    if 'rxn' in reaction.id:
        tmp = {str(k.id):v for k,v in reaction.metabolites.items()}
        if any(x in tmp for x in atp_sim_list):
            atp_sim_reaction.append(reaction.id)
print(len(atp_sim_reaction))
print(atp_sim_reaction)

co2_reaction = []

for reaction in yeast8U_mod.reactions:
    if 'rxn' in reaction.id:
        tmp = {str(k.id):v for k,v in reaction.metabolites.items()}
        if 's_0456' in tmp and tmp['s_0456']<0:
            co2_reaction.append(reaction.id)

print(len(co2_reaction))
print(co2_reaction)

o2_reaction = []

for reaction in yeast8U_mod.reactions:
    if 'rxn' in reaction.id:
        tmp = {str(k.id):v for k,v in reaction.metabolites.items()}
        if 's_1275' in tmp and tmp['s_1275']>0:
            o2_reaction.append(reaction.id)

print(len(o2_reaction))
print(o2_reaction)

nadph_reaction = []
# nadph nadh fadh2

for reaction in yeast8U_mod.reactions:
    if 'rxn' in reaction.id:
        tmp = {str(k.id):v for k,v in reaction.metabolites.items()}
        if 's_1212' in tmp and tmp['s_1212']>0:  #NADPH
            nadph_reaction.append(reaction.id)

for reaction in yeast8U_mod.reactions:
    if 'rxn' in reaction.id:
        tmp = {str(k.id):v for k,v in reaction.metabolites.items()}
        if 's_1203' in tmp and tmp['s_1203']>0:   #NADH
            nadph_reaction.append(reaction.id)

for reaction in yeast8U_mod.reactions:
    if 'rxn' in reaction.id:
        tmp = {str(k.id):v for k,v in reaction.metabolites.items()}
        if 's_0689' in tmp and tmp['s_0689']>0:  #FADH2
            nadph_reaction.append(reaction.id)

print(len(nadph_reaction))
print(nadph_reaction)


# ['s_1360','s_0075','s_4092','s_1464'] 

yeast8_atp = []

for reaction in yeast8U_mod.reactions:
    if 'rxn' in reaction.id:
        tmp = {str(k.id):v for k,v in reaction.metabolites.items()}
        if 's_1360' in tmp and tmp['s_1360']>0:   #phosphoenolpyruvate
            yeast8_atp.append(reaction.id)

for reaction in yeast8U_mod.reactions:
    if 'rxn' in reaction.id:
        tmp = {str(k.id):v for k,v in reaction.metabolites.items()}
        if 's_0075' in tmp and tmp['s_0075']>0:   #1,3-bisphospho-D-glycerate
            yeast8_atp.append(reaction.id)

for reaction in yeast8U_mod.reactions:
    if 'rxn' in reaction.id:
        tmp = {str(k.id):v for k,v in reaction.metabolites.items()}
        if 's_4092' in tmp and tmp['s_4092']>0:  #N(omega)-phospho-L-arginine
            yeast8_atp.append(reaction.id)

for reaction in yeast8U_mod.reactions:
    if 'rxn' in reaction.id:
        tmp = {str(k.id):v for k,v in reaction.metabolites.items()}
        if 's_1464' in tmp and tmp['s_1464']>0:  #succinyl-CoA
            yeast8_atp.append(reaction.id)


print(len(yeast8_atp))
print(yeast8_atp)

else_reaction = []

for reaction in yeast8U_mod.reactions:
    if 'rxn' in reaction.id:
        tmp = {str(k.id):v for k,v in reaction.metabolites.items()}
        if 's_0764' in tmp and tmp['s_0764']>0:     #glycerol
            else_reaction.append(reaction.id)

for reaction in yeast8U_mod.reactions:
    if 'rxn' in reaction.id:
        tmp = {str(k.id):v for k,v in reaction.metabolites.items()}
        if 's_0362' in tmp and tmp['s_0362']<0:   #acetate
            else_reaction.append(reaction.id)

for reaction in yeast8U_mod.reactions:
    if 'rxn' in reaction.id:
        tmp = {str(k.id):v for k,v in reaction.metabolites.items()}
        if 's_0140' in tmp and tmp['s_0140']>0:  #2,3-bisphospho-D-glyceric acid
            else_reaction.append(reaction.id)

for reaction in yeast8U_mod.reactions:
    if 'rxn' in reaction.id:
        tmp = {str(k.id):v for k,v in reaction.metabolites.items()}
        if 's_0979' in tmp and tmp['s_0979']<0:  #L-citrulline
            else_reaction.append(reaction.id)

for reaction in yeast8U_mod.reactions:
    if 'rxn' in reaction.id:
        tmp = {str(k.id):v for k,v in reaction.metabolites.items()}
        if 's_0373' in tmp and tmp['s_0373']>0:  #acetyl-CoA
            else_reaction.append(reaction.id)

print(len(else_reaction))
print(else_reaction)

# Calculate growth after preliminary screening

reaction_id = []
for i in yeast8U_mod.reactions:
    if 'rxn' in i.id:
        i.bounds = (0,0)
        reaction_id.append(i.id)

reaction_id = [x for x in reaction_id if x not in biomass_up_reaction]
reaction_id = [x for x in reaction_id if x not in atp_reaction]
reaction_id = [x for x in reaction_id if x not in atp_sim_reaction]
reaction_id = [x for x in reaction_id if x not in co2_reaction]
reaction_id = [x for x in reaction_id if x not in o2_reaction]
reaction_id = [x for x in reaction_id if x not in nadph_reaction]
reaction_id = [x for x in reaction_id if x not in yeast8_atp]
reaction_id = [x for x in reaction_id if x not in else_reaction]


# for seed_num in range(20,40):
#     random.seed(seed_num)
#     random.shuffle(reaction_id)
#     with yeast8U_mod:
#         error_reaction_4 = []
#         for i in tqdm(reaction_id):
#             reaction = yeast8U_mod.reactions.get_by_id(i)
#             if 'rxn' in reaction.id:
#                 reaction.bounds = (0,1000)
#                 yeast8U_mod.solver = 'gurobi'
#                 yeast8U_mod.optimize()
#                 if yeast8U_mod.reactions.get_by_id('r_2111').flux > 0.1:
#                     reaction.bounds = (0,0)
#                     error_reaction_4.append(reaction.id)
#         print(seed_num,len(error_reaction_4))
#     # Store the list in a binary file
#     filename = error_reaction_path + 'error_reaction_random'+str(seed_num)+'.pkl'
#     with open(filename, 'wb') as file:
#         pickle.dump(error_reaction_4, file)

def process_seed(args):
    seed_num, reaction_id, yeast8U_mod, error_reaction_path = args
    random.seed(seed_num)
    random.shuffle(reaction_id)
    with yeast8U_mod:
        error_reaction_4 = []
        for i in tqdm(reaction_id):
            reaction = yeast8U_mod.reactions.get_by_id(i)
            if 'rxn' in reaction.id:
                reaction.bounds = (0, 1000)
                yeast8U_mod.solver = 'gurobi'
                yeast8U_mod.optimize()
                if yeast8U_mod.reactions.get_by_id('r_2111').flux > 0.09:
                    reaction.bounds = (0, 0)
                    error_reaction_4.append(reaction.id)
        print(seed_num, len(error_reaction_4))
    # Store the list in a binary file
    filename = error_reaction_path + 'error_reaction_random' + str(seed_num) + '.pkl'
    if not os.path.exists(error_reaction_path):
        os.makedirs(error_reaction_path)
    with open(filename, 'wb') as file:
        pickle.dump(error_reaction_4, file)



if __name__ == '__main__':
    with Pool(20) as pool:
        args = [(seed_num, reaction_id, yeast8U_mod, error_reaction_path) for seed_num in range(20, 40)]
        results = list(tqdm(pool.imap(process_seed, args), total=len(args)))

# Select and delete the pkl file with the fewest reactions

# error_reaction_path = "../../Results/not_lipid/top50_0.3/error_reaction_random/"
files = os.listdir(error_reaction_path)

# Find the file with the minimum error_reaction
min_error_reaction = 1000
min_error_reaction_file = None

for file in files:
    file_path = os.path.join(error_reaction_path, file)
    with open(file_path, 'rb') as f:
        error_reaction_data = pickle.load(f)
        if len(error_reaction_data) < min_error_reaction:
            min_error_reaction = len(error_reaction_data)
            min_error_reaction_file = file  # Store the file name

print(min_error_reaction_file)


file_path = os.path.join(error_reaction_path, min_error_reaction_file)
with open(file_path, 'rb') as f:
    error_reaction_data = pickle.load(f)
print(f"Loaded data length: {len(error_reaction_data)}")


error_reaction_data

del_reaction_lst = biomass_up_reaction + atp_reaction + atp_sim_reaction + co2_reaction + o2_reaction + nadph_reaction + yeast8_atp + else_reaction + error_reaction_data
print(len(del_reaction_lst))
del_reaction_lst = list(set(del_reaction_lst))
print(len(del_reaction_lst))

del_reaction_lst

# filename = '../Data/analysis/error_reaction_all.pkl'
with open(error_reaction_all_path, 'wb') as file:
    pickle.dump(del_reaction_lst, file)

# with open(error_reaction_all_path, 'rb') as file:
#     del_reaction_lst_ = pickle.load(file)
# del_reaction_lst_

rxndb_total_info_to_model = pd.read_csv(rxndb_total_info_to_model_path)
print(rxndb_total_info_to_model.shape)
rxndb_total_info_to_model = rxndb_total_info_to_model[~rxndb_total_info_to_model['NO'].isin(del_reaction_lst)]
print(rxndb_total_info_to_model.shape)
rxndb_total_info_to_model.to_csv(rxndb_total_info_to_model_del_reaction_path, index=False)

get_yeast8U(metabolites_info_to_GEM_path,rxndb_total_info_to_model_del_reaction_path,yeast8_reaction_in_rxndb_json,yeast870_path,yeast8U_del_path,metinfoID_2_YMDBID=None)

yeast8U_del = cobra.io.load_yaml_model(yeast8U_del_path)
yeast8U_del.summary()

# for reaction_id in tqdm(del_reaction_lst):
#     reaction = yeast8U_mod.reactions.get_by_id(reaction_id)
#     yeast8U_mod.remove_reactions([reaction])

# yeast8U_del_path = '../Data/model/yeast8U_Final_mod_del.yml'

# cobra.io.save_yaml_model(yeast8U_mod,yeast8U_del_path)
