==> Code/retrosynthesis/06-4.get_YeastMetaTwin.ipynb <==
import pandas as pd
import numpy as np
import json
import pandas as pd
from cobra import Model, Reaction, Metabolite
from cobra.flux_analysis import flux_variability_analysis
import pickle
from tqdm import tqdm 
from ast import literal_eval
import cobra
import matplotlib.pyplot as plt
import numpy as np
from multiprocessing import Pool
from tqdm import tqdm
from functools import partial 
import matplotlib.pyplot as plt
import numpy as np
from matplotlib.gridspec import GridSpec
from matplotlib.ticker import MaxNLocator, FormatStrFormatter

import gurobipy
gurobipy.setParam('OutputFlag', 0)
import sys
# sys.path.append('../Script/')
from common import *

# input and output

###input
# yeast870_path =  '../../Data/model/yeast-GEM.yml'
yeast8U_del_path = '../../Data/model/yeast8U_del_re.yml'
yeast870_path =  '../../Data/model/yeast-GEM902.yml'
YMDB_total_smiles_path = '../../Data/ymdb/ymdb_final_result.xlsx'
ess_del_reaction_df_file = '../../Results/analysis/ess_del_reaction_df_re.csv'
biolog_del_reaction_df_file = '../../Results/analysis/Biolog_del_reaction_re.csv'

cut_off = 0.3
num = 50
num_process = 50
cut_off_path = f'../../Results/not_lipid/top{num}_{cut_off}_re/'
metabolites_info_to_GEM_path = cut_off_path + f'metabolites_info_to_GEM_top{num}_{cut_off}.csv'
# rxndb_total_info_to_model_path = cut_off_path + f'rxndb_total_info_to_model_top{num}_{cut_off}.csv'
yeast8_reaction_in_rxndb_json = cut_off_path + f'yeast8_reaction_in_rxndb_top{num}_{cut_off}.json'
rxndb_total_info_to_model_del_reaction_path = '../../Results/not_lipid/top50_0.3_re/rxndb_total_info_to_model_del_reaction_top50_0.3.csv'

###output
rxndb_total_info_to_model_del2_reaction_path = '../../Results/not_lipid/top50_0.3_re/rxndb_total_info_to_model_del2_reaction_top50_0.3.csv'
metabolites_info_to_GEM_del_path = '../../Results/not_lipid/top50_0.3_re/metabolites_info_to_GEM_del_top50_0.3.csv'
yeast8U_del2_path = '../../Data/model/Yeast-MetaTwin.yml'

del_reaction_lst = []

ess_del_reaction_df = pd.read_csv(ess_del_reaction_df_file,index_col=None)
ess_del_reaction_df['reaction'] = ess_del_reaction_df['reaction'].apply(lambda x : literal_eval(x))
ess_del_reaction_df = ess_del_reaction_df[ess_del_reaction_df['num']<20]

for i in ess_del_reaction_df['reaction'].to_list():
    del_reaction_lst += i

print(len(del_reaction_lst))
del_reaction_lst = list(set(del_reaction_lst))
print(len(del_reaction_lst))

print('==============================')

biolog_del_reaction_df = pd.read_csv(biolog_del_reaction_df_file,index_col=None)
biolog_del_reaction_df['reaction'] = biolog_del_reaction_df['reaction'].apply(lambda x : literal_eval(x))
biolog_del_reaction_df = biolog_del_reaction_df[biolog_del_reaction_df['num']<20]

for i in biolog_del_reaction_df['reaction'].to_list():
    del_reaction_lst += i

print(len(del_reaction_lst))
del_reaction_lst = list(set(del_reaction_lst))
print(len(del_reaction_lst))

rxndb_total_info_to_model = pd.read_csv(rxndb_total_info_to_model_del_reaction_path)
print(rxndb_total_info_to_model.shape)
rxndb_total_info_to_model = rxndb_total_info_to_model[~rxndb_total_info_to_model['NO'].isin(del_reaction_lst)]
print(rxndb_total_info_to_model.shape)
rxndb_total_info_to_model.to_csv(rxndb_total_info_to_model_del2_reaction_path, index=False)

# met_info add YMDBID and YMDBname

met_info = pd.read_csv(metabolites_info_to_GEM_path)
met_info = met_info[~met_info['ID'].str.contains('s_')]
YMDB_Data = pd.read_excel(YMDB_total_smiles_path)
YMDB_Data = YMDB_Data[YMDB_Data['in_model']==0]

def process_metinfoID(metinfoID, met_info, YMDB_Data):
    metinfoID_smiles = met_info[met_info['ID'] == metinfoID]['new_met_smiles'].to_list()[0]
    YMDB_ID = []
    for index, row in YMDB_Data.iterrows():
        if compare_smiles_inchikey(metinfoID_smiles, row['SMILES']):
            YMDB_ID.append(row['ID'])
    return metinfoID, YMDB_ID

metinfoID_2_YMDBID = {}
metinfoID_list = list(set(met_info['ID'].to_list()))
process_partial = partial(process_metinfoID, met_info=met_info, YMDB_Data=YMDB_Data)
with mp.Pool(60) as pool:
    results = list(tqdm(pool.imap_unordered(process_partial, metinfoID_list), total=len(metinfoID_list)))
for metinfoID, YMDB_ID in results:
    metinfoID_2_YMDBID[metinfoID] = YMDB_ID

# reset rxnID

rxndb_to_model = pd.read_csv(rxndb_total_info_to_model_del2_reaction_path)
rxndb_to_model['equation_dict'] = rxndb_to_model['equation_dict'].apply(lambda x:literal_eval(x))
rxndb_to_model = rxndb_to_model[~rxndb_to_model['equation'].str.startswith(' =>')]
rxndb_to_model = rxndb_to_model[~rxndb_to_model['equation'].str.endswith(' => ')]
rxndb_to_model = rxndb_to_model.drop_duplicates(['equation'], keep='first')
with open(yeast8_reaction_in_rxndb_json, 'r') as file:
    yeast8_reaction_in_rxndb_dict = json.load(file)
yeast8_reaction_in_rxndb = list(set([x for lst in yeast8_reaction_in_rxndb_dict.values() for x in lst]))
rxndb_to_model = rxndb_to_model[~rxndb_to_model['NO'].isin(yeast8_reaction_in_rxndb)]
rxndb_to_model = rxndb_to_model.reset_index(drop=True)
rxndb_to_model['NO'] = 'rxn' + (rxndb_to_model.index + 1).astype(str)
mask = rxndb_to_model['GPR'] == 's_1025 + s_0794 => s_0456 + s_4299'
rxndb_to_model.loc[mask & ~rxndb_to_model['GPR'].str.contains('YKL184W'), 'GPR'] = (
    rxndb_to_model.loc[mask & ~rxndb_to_model['GPR'].str.contains('YKL184W'), 'GPR'] 
    + ' or (YKL184W)'
)
rxndb_to_model.to_csv(rxndb_total_info_to_model_del2_reaction_path,index=None)
rxndb_to_model.head()

get_yeast8U_step4(metabolites_info_to_GEM_path,rxndb_total_info_to_model_del2_reaction_path,yeast870_path,yeast8U_del2_path,metinfoID_2_YMDBID)

yeast8 = cobra.io.load_yaml_model(yeast870_path)
yeast8.summary()

yeast8U_mod = cobra.io.load_yaml_model(yeast8U_del2_path)

yeast8U_mod.summary()

rxndb_to_model = pd.read_csv(rxndb_total_info_to_model_del2_reaction_path)
metabolites_info_to_GEM = pd.read_csv(metabolites_info_to_GEM_path)
met_id_lst = []
rxndb_to_model['equation_dict'] = rxndb_to_model['equation_dict'].apply(lambda x:literal_eval(x))
for i in range(len(rxndb_to_model)):
    met_id_lst += list(rxndb_to_model['equation_dict'].iloc[i].keys())

met_id_lst = list(set(met_id_lst))
print(metabolites_info_to_GEM.shape)
metabolites_info_to_GEM_del = metabolites_info_to_GEM[metabolites_info_to_GEM['ID'].isin(met_id_lst)]
print(metabolites_info_to_GEM_del.shape)
metabolites_info_to_GEM_del.to_csv(metabolites_info_to_GEM_del_path, index=False)
